+++ main.go
@@ -1,14 +1,34 @@
 package main
 
 import (
+	"time"
+
+	l "dhc-server/logging"
+
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 )
 
 var context = DatabaseContext{DatabaseName: "dhc.db"}
 
+func SetupLogger(logger *l.Logger) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		t := time.Now()
+
+		c.Next()
+		latency := time.Since(t)
+
+		status := c.Writer.Status()
+
+		logger.Info("response",
+			l.LogArg{Key: "latency", Value: latency},
+			l.LogArg{Key: "status_code", Value: status},
+		)
+	}
+}
+
 func setupRouter() *gin.Engine {
-	logger, err := NewLogger()
+	logger, err := l.NewLogger()
 	if err != nil {
 		panic(err)
 	}
